requirements.txt
configparser==4.0.2
cryptography==1.2.3
cycler==0.10.0
Cython==0.29.14
h5py==2.10.0
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-resnet==0.1.0
keras-retinanet==0.5.1
lxml==3.5.0
Markdown==3.1.1
matplotlib==3.0.3
numpy==1.17.4
opencv-python==4.1.2.30
Pillow==3.1.2
progressbar2==3.47.0
pycocotools==2.0
pyparsing==2.0.3
six==1.10.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
